Speed up cache access
Make functions in pyfaf.run access cache directly instead of executing
faf-cache script through pipe. Simple unit test provided.

Note that when reading/writing binary cache values, the whole content is
loaded into memory (it was processed in blocks previously). The
performance impact is not obvious, so it may be neccessary to change
this if it turns out to be a problem.

Closes trac#47.