pyfaf: core problemtype: handle '??' in last frame
function_name = '??' was handled differently for the last frame
and the rest. For the rest of the frames it is replaced
with 'anonymous function' in case there are ?? but not for
last frame where it's only replaced with 'anonymous function'
in case it is missing completely.

This patch unifies this behavior.

Signed-off-by: Richard Marko <rmarko@fedoraproject.org>