pyfaf: better reuse old problems in create-problems
Previously an old problem had to match in at least a half of
its reports to be reused. This methods reuses the old problem
if at least one report matches. Also properly connects the best
matches together.
This should reduce the number of empty problems left after a run
of create-problems.