pyfaf: add retrace --max-fail-count
This option allows to stop retracing a symbol after a given
count of failed runs. In the production DB there are many symbols
that fail over and over and eat the majority of processing time.
The action should still be run occassionaly without this option
to to retry retracing of all symbols.