pyfaf: pickle saved report ids only once in pull-reports
The pull-reports action uses pickle to save the set of already pulled reports
to avoid pulling duplicated. Since the pickle can grow in size a lot, it can
take a couple of seconds to save to the disk. This change makes the pickle to
be saved only once after pulling all the reports instead of saving it after
every single pull. This significantly improves the speed of pulling.