pyfaf: improve bucketing efficiency
Previously the size of buckets in progress was determined by the sum of
sizes of thread sets in a temporary group. These temporary groups were then
unioned together to form one set. If many of the thread sets in a group were
similar, the union could decrease the size of the set below the maximum which
is inefficient.  Now the union is computed incrementaly, and the size
determined from directly from the current union, so this cannot happen.