webfaf: Added method items, get_hash, get_avg_count and parameter want_object
Method items: In POST accept a JSON with bthashes and report for every hash.
This method return json which contain a list of reports. Method was created
because we need to get a lot of reports and we don't want send a huge amount
of requests.

Method get_hash: This method was created because we need download all available
bthashes which was created on FAF. Get_hash can accept 4 parametrs:
os - Select only hashes which have reports for specific OS (Fedora, RHEL, CentOS)
release - This parameter requires to have specified parameter OS. By this parameter
you select only specific release of OS.
since and to - This parameters selected hashes which containt at least one report
have last occurrence between this dates.

To method Item was added parameter want_object. It's because method items
call this method in cycle and need return objects.
Also this method was extended, now we return maintainers of component and if was report
probably fixed.

Method get_avg_count: return average occurrence of report per month

Signed-off-by: Patrik Helia <phelia@redhat.com>