repos: drop dash from RpmMetadata's name
I am sorry for this follow up commit but I was under the impression that
RpmMetadata.name will be used as plugin name. This is not true and I
have to admin that I don't fully understand the purpose of that member.

The RpmMetadata's plugin name is 'rpmmetadata' (as generated by
common.load_pluing_types) and I have decided to set the
RpmMetadata.name class member to the same value in order to avoid
confusion and possible problems with inconsistencies (some code could
start using that member and try to match it against the generated name).

Signed-off-by: Jakub Filak <jfilak@redhat.com>