actions: teach reposync to process subset of packages
On some systems with limited resources we need to split the process of
repository synchronizations into several steps. We can either process
the repository packages by parts or we can have more repositories each
containing subset of the original repository. The former approach seems
to be more general because we do not always have the repositories under
our control.

This commit introduces a new command line option 'name-prefix' in the
reposync action. Passing the value in command line arguments causes that
only packages whose name converted to lower string starts with the
prefix converted to lower string will be processed.

It is not sure if there will be other use case than passing all alphabet
letters and digits, though.

Signed-off-by: Jakub Filak <jfilak@redhat.com>