retrace: correct formatting of address to hex
int and long are unified but even if you cast (sys.maxint + 1) to int the
output of hex has 'L' suffix.

In the commit ba5deb189113c135836057803a729ca2f7d7ccef, I did not
realize the addr2line function accepts address as a number and converts
it to a string on its own. So that commit was a mistake and the right
place where the _correct_ format should be is the function.

Signed-off-by: Jakub Filak <jfilak@redhat.com>