storage: add a list like wrapper for Query.yield_per
For the cases where we just need to iterate over all items and we want
to print valuable progress messages, we have to first load all items into
a list and then iterate over the list.

In some hostile environments where we have limited memory resources, we
cannot load all items into a list, so we need to use yeild_per(). This
approach implies, we first must call count() and then yield_per().

This commit introduces a wrapper class that ensures that we do not need to
add a special casing when we need to use yield_per(). We can call len()
on the query results and we can also use it in the for cycle as we are
used to.

Signed-off-by: Jakub Filak <jfilak@redhat.com>