problemtypes: support yield_per() when reading symbol sources
It is not always possible to load all symbol sources that need to be
retraced into memory.

I have decided to add a new abstract *protected* method which should
return a pure Query or None if the problem type does not support
retracing. The method is called by the already existing method in the
problem type super class. The public method will do the common work,
i.e. add limit for max failures and yield results.

Signed-off-by: Jakub Filak <jfilak@redhat.com>