pyfaf: optimize symbol source queries for yield_per
This commit moves the join part of those queries into a subquery in
order to enable the count() method to return accurate results.

Basically, we want to select all symbol sources that match a certain
rule and the join was causing that query results included some symbol
sources multiple times. SQLAlchemy handles this problem on its own, so
when we load all symbol sources into a list len() returns the correct
result.

Signed-off-by: Jakub Filak <jfilak@redhat.com>