Webfaf: Integrate solutions, solution filter in report and problem list
Added filter to WEBUI for showing reports which have certainty on 100
this mean the report have a solution.

If report have solution, then solution note is show in detail of report
In report detail WEBUI added support for more solution then one.

Added filter to WEBUI for showing reports which have certainty on 100
this mean the report have a solution.

If problem have a report with solution then this report is by default hide
in report's summary (Added optional filter for enable showing this problems).
If problem have a report with solution, then is solution showed in problem's
detail

Signed-off-by: Patrik Helia <phelia@redhat.com>