Introduce a way of assigning release to a repository
There are currently two kinds of repositories:
    1) Repos assigned with specific operating system (so called parametrized
            repositories)
    2) Pure repos without relations

This commit introduces third kind of repository - repo assigned to operating
system and release. This repo is similar to the first kind, but its url does
not contain any parameters (dollar signs).

To create this kind of repository a 'repoassign' action must have been edited,
since it was not capable of testing releases for their presence. Now each
parameter is not only tested if it is operating system or architecture, but
also is parsed and tested for being operating system with release. (operating
system must be separated by space from release, see example)
Therefore it is possible to pass arguments containing operating system with
release.
   Example: `faf repoassign repo_name "Fedora 24" "rhel 6" x86_64`

When `repoassign` action could parse releases, it needed to save them somewhere.
For this purpose a new table must have been introduced. Since repositories of
the first type have table 'OpSysRepo', table following the same logic was
added - 'OpSysReleaseRepo'. This also meant adding reference to OpSysRelease
into Repo table. This action created error, therefore the OpSysRelease table
got to moved lower in the code.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>