Teach reposync to assign builds to operating system and release
It would be convenient to know for each build to which operating system release
it belongs. Knowing this it would be possible to perform a efficient garbage
collection of packages/builds of EOL releases.

To make this possible `reposync` action must have information about operating
system and release. These information can be found in repository, if they were
assigned beforehand with action `repoassign`.

To be able to store these information a link table must be introduced. There
is need to store connection between Build, operating system, release and
architecture.

In the link table, architecture is the one that was assigned to repository
with action repoassign. Therefore each repository must be assigned with
at least one architecture. Thanks to this link table, more than one architecture
can be assigned.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>