tests: Add test for alembic migration tool
Introduce test for testing if there is only one head for alembic tree.
Creating new revision automatically sets down version to the current head. But
if in the time of creating author does not have current head, or the head
changes in the time between creating revision and merging the code, it can
create new head, which is not intended.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>