storage: Fix inadequate check for semver validity
There exists build 'ginga-2.5.20161005204600-1.fc25.noarch.rpm'. When
this build was reposynced, exception from database (stating that semver
        is bigger that 31 bits) caused the whole reposync to exit.

I've decided instead of putting code into try-except block to check if
the parts are really smaller than 31 bits. If not, then the semver
cannot be marked as correct and fitting is needed.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>