docker: Introduce FAF docker image
Obsoletes abrt/faf-docker repository.

Why have I created a new docker image?

Firstly, because I didn't know there was one.
Secondly, the image needed some rewriting and rethinking.

So why is it directly in FAF repo?
Because it is the correct place. We do not need another repository for
so little data. And also, nobody misses it ever again:)

How does it differ?
The biggest difference is that it builds FAF from currently checked out
git branch. It makes it perfect for development. It makes however, the image a
bit bigger (there are some build-requires). Therefore there is another
Dockerfile, which builds it from copr repository. This is production
image.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>