opsys: Do not require packages in uReport
This is second step in effort to make it possible to use FAF as common
bug catcher.
The main disadvantage is that uReport must contain packages. This commit
enables reports with no packages. However the check for at least one
package was kept, but only when processing unpackaged executables is
enabled.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>