docker: Fix permissions for openshift user
Openshift Origin runs containers using user ID >= 1000000000.

The user is part of the root group. To allow this user to use/write
to faf files/directories the permissions on the files/directories
should be altered so they are root group readable/writable/executable.

The user ID of the container is generated dynamically, that means there is no
associated entry in /etc/passwd. This causes problems with faf web.
To fix this problem we add a new entry to the /etc/passwd as part of the run script.

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>