webfaf: Limit the number of exception mails
Limits the rate at which the messages are passed to the SMTP handler.
It uses a token bucket based algorithm to filter the messages.

Ex.1:
* rate=1 timeframe=30 burst=1
After a mail is sent there is a 30s period until another one can be sent.

*Ex.2: rate=1 timeframe=60 burst=5
Allows to send 5 mails in burst before the filtering kicks in. After that
the number of mails that can be sent starts recovering at the rate of 1 mail
per 60s.

There is also possibility to filter the mails based on the substring
in the exception message.

More info at https://github.com/wkeeling/ratelimitingfilter

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>