py3 compatibility: webfaf str vs bytes fixes
In Python 2, the 'str' type can be used for both text and binary data.
It does not work in Python 3 where must be distinguished which APIs
take text and which take binary. There must be also considered that
Python 2 text APIs can work with 'unicode' and APIs that work
with binary data, must work with the bytes type from Python 3.

Signed-off-by: Jan Beran <jberan@redhat.com>