create-problems: Use set instead of HashableSet
Sets are implemented in a way, which doesn't allow mutable objects, but
sets are mutable. Hashable objects must be immutable. The problem was
solved by 'HashableSet()' object.

Python 3 contains 'frozenset()' type which is the immutable variant
of 'set()' and is hashable. Python 2 is not so strict as Python 3,
and allows to modify the hashable set, which is impossible to do
in Python 3. The patch solves the issue.

Signed-off-by: Jan Beran <jberan@redhat.com>