schema: Create Message schema for FAF
Adds faf-schema subpackage that defines message schema for the fedora-messaging
infrastructure. The fedora-messaging library is designed to be a replacement for
the fedmsg library.

Message schema should be treated as immutable. Once defined,
they should not be altered. Instead, define a new schema class,
mark the old one as deprecated, and remove it after an appropriate transition period.[1]

Docs:
[1] https://fedora-messaging.readthedocs.io/en/latest/messages.html#schema-are-immutable
https://fedora-messaging.readthedocs.io/en/latest/publishing.html
https://fedora-messaging.readthedocs.io/en/latest/tutorial/conversion.html

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>