dumpdirs: Modify a route to accept uploads of dumpdirs
If a dumpdir is sent to 'retrace.fp.org/faf/dumpdirs/new/xyz.tar' server returns '301 MOVED PERMANENTLY'
and a request is redirected to 'retrace.fp.org/faf/dumpdirs/new/xyz.tar/' (see [1]).

This causes Flask to not process the upload instead return FormDataRoutingRedirect error.

When 'strict_slashes' are disabled the redirect is not triggered.

[1] http://werkzeug.pocoo.org/docs/0.14/routing/#rule-format

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>