pyfaf, webfaf: Display version number in interface
Display version information in web interface and add a flag
(-V/--verbose) for faf CLI to print the same.

The number is taken from faf-version and stored in the pyfaf module
during the build process. It is then pulled by the CLI and webfaf.

Also, call sys.exit() in bin/faf rather than exit() which is implicitly
imported from the site module, as it may lead to problems.