tests: rpm_metadata: Avoid races
Currently, the socket for the HTTP server is created when the thread
is started, requiring the following code to sleep for a bit. Even with
that, it seems that the test fails on COPR due to some requests being
refused. Not sure what exactly is going on, but smells like a timing
issue as well. This commit rewrites the test in a way that ensure that
the socket starts listening way before we start serving requests.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>