create_problems: Use multiple threads to process db_reports
We can save some time if we use more workers (threads) to process the db_reports
concurrently.

The communication with the DB is slowing down the db_report processing.
So while some of the workers are loading reports from the DB, other workers can start
processing the loaded reports.

The number of workers is limited by 'pool_size' and 'max_overflow' settings of connection engine.
(Default is 5 for pool_size and 10 for max_overflow).

Default is 4 threads.

faf create-problems -w 8
faf create-problems --max-workers 8

Signed-off-by: Martin Kutlak <mkutlak@redhat.com>