webfaf: Fix serialization of reports with solutions
Currently, a dictionary is being passed to jsonify() from Flask that
includes an array of Solution objects (if any exist). The result is a
TypeError, because Python classes are not JSON-serializable, hence the
need to take __dict__ first. Moreover, the “since” attribute of type
datetime.datetime needs to be stringified, because it’s also not
JSON-serializable.