Rework container build
Currently, the “main” image is built by using RPMs from Copr, which
hurts reproducibility and automation, given that we have to build
everything in Copr first and only then build the container image. This
commit leaves only one Dockerfile file for production and development.
The image is built in multiple stages to prevent unwanted artifacts
staying around:

    1. Build FAF RPMs
    2. Upgrade system and install FAF RPMs
    3. Perform post-installation setup

Additionally, this reworks the Makefile shorthands: the names for
images, pods and volumes were changed to drop the redundant suffixes
(-image, -volume), the targets that depend on files existing now
actually do depend on them and common options were moved out to
variables.