webfaf: Refactor report history charts
Move the JavaScript plotting procedures into static and call them on
document load for designated elements.

Modify the aesthetics to a less cluttered format, which should also be
more readable.

The format of histories is also changed, which will break those who
rely on the JSON format of the report page.