pyfaf: mantisbt: Don’t compare ObjectRef to string
Most likely a thinko, in that, for determining whether the report is
duplicate, the code compares an ObjectRef from Zeep with a textual
representation of the resolution. The code above converts the
(localized!!!) name of the resolution to an internal representation (one
that matches what Bugzilla has).

This commit fixes the aforementioned thinko.