README: Add instructions for container deployments
Since automated builds in Docker Hub result in readmes being wiped and
replaced with upstream ones, we need to provide the instructions there.

This commit moves the bulk of the documentation from container/README.md
and reformats it slightly for brevity. Additionally, what’s left of
container/README.md gets the same treatment.