repos: dnf: Correctly handle debug packages
5508c07b919a159e56acaad0bd192e02297152b8 completely ignored the problem
of base package names, resulting in debuginfo and debugsource packages
having distinct builds in the database and breaking retracing. While
there is no longer a property in DNF that spits out the base name, we
can use the class attributes and strip the suffixes as needed. This
commit does just that.

Fixes https://github.com/abrt/faf/issues/919