pyfaf: faf_rpm: Don’t parse EVRs with non-numeric epochs
In RPM versions prior to 4.14 (but not 4.9, there was a bit of a
back-and-forth), unexpanded macros can go unnoticed in dependency
declarations, leading to reposync throwing a fit when trying to parse
the EVR of any related package to the one being processed. Any
dependency that includes %epoch without an epoch set will have the macro
unexpanded, leaving us with garbage, since epochs are strictly numeric.

This commit adds an additional check in parse_evr() to validate the
epoch, bailing out in case of trouble.

Related RPM commits:
https://github.com/rpm-software-management/rpm/commit/507f21f6bb4bf7029a0bca255cfe4aae3361f358
https://github.com/rpm-software-management/rpm/commit/f5bab7c054de3607d43dcc5ee8a04c2e3cd46926

Closes https://github.com/abrt/faf/issues/927