pyfaf: actions: reposync: Add --no-cache argument
Even with the ability to configure the cache lifetime, it can be useful
to have a one-shot way to go around it. The change is invasive, however:
loading the packages in __init__() makes everyone’s lives that much more
difficult, as it is an expensive and blocking operation, as well as one
that prevents any post-construction changes to the configuration. What’s
even worse is that it’s called *for* *each* repo in a loop, which wastes
even more time.

Closes https://github.com/abrt/faf/issues/928