webfaf: reports: Fix non-Bugzilla bug association
Currently, the code assumes that any new report association with an
external tracker bug will be strictly Bugzilla, which is not the only
tracker FAF supports. This commit fixes that by adding synonyms to
report tables in order to achieve some kind of genericity along with
dynamic inspection of selected bug tracker on POST to
/<report>/associate_bz.

The same issue applies to disassociation, where the situation is a bit
more complex, in that we don’t know what tracker we’re using beforehand,
so this commit adds a new database query method that iterates all of
them.

Closes https://github.com/abrt/faf/issues/924