celery: Fix invocation with global options
Celery 5.0 removed support of global options after sub-command.

Changes still work in older Celery versions (4.x is used in CentOS 8).

The error message from Celery 5.0:
You are using `-A` as an option of the beat sub-command:
celery beat -A celeryapp <...>
The support for this usage was removed in Celery 5.0. Instead you should use `-A` as a global option:
celery -A celeryapp beat <...>

See for more info:
https://docs.celeryproject.org/en/stable/whatsnew-5.0.html#step-1-adjust-your-command-line-invocation