celery_tasks: Remove future-dependent annotation
Fallout from https://github.com/abrt/faf/pull/967. Type annotation
without futures causes a crash when accessing the FAF web interface
because the class in the annotation is only defined later in the file.
Moving the code around causes the same issue with a different class so
I'm just dropping the annotation.

Signed-off-by: Michal Fabik <mfabik@redhat.com>