faf_rpm: Skip misformatted Provides in RPM headers
SQLAlchemy's SAVEPOINT transactions seem to cause unanticipated issues
when used in conjunction with autocommit mode. Namely, trying to call
`Session.begin_nested()` on its own raises an exception due to the way
transactions are handled in autocommit mode.

This bug was introduced 334ce06f, but was missed in tests because they
are run with `autocommit=False` while production code is run with
`autocommit=True`.

Instead of working with (sub)transactions and in order to avoid the
fluff around SQLAlchemy's autocommit magic, we just skip the records in
question. When we encounter a misformatted 'Provides' in the RPM header,
instead of trying to roll back the (sub)transaction, warn about the
matter, skip the item and continue with the remainder of the header.

Fixes #958