Add hackish support for JIRA
FAF is quite coupled with Bugzilla. It can be extended, but for example
tables in Postgres are Bugzilla-specific. So, in order to add
a new bug tracker backend, we would need to introduce new tables
as well.

It's also not clear why is FAF trying to "mirror" so much data
from Bugzilla... The code base is old and full of terrors...

Signed-off-by: Michal Srb <michal@redhat.com>